chore: create simu node classes
