+++ substrafl/experiment.py
@@ -2,13 +2,11 @@ import copy
 import datetime
 import json
 import logging
-import types
 import uuid
 from functools import reduce
 from operator import add
 from pathlib import Path
 from platform import python_version
-from typing import Any
 from typing import Dict
 from typing import List
 from typing import Optional
@@ -30,11 +28,9 @@ from substrafl.exceptions import UnsupportedClientBackendTypeError
 from substrafl.nodes.aggregation_node import AggregationNode
 from substrafl.nodes.schemas import OperationKey
 from substrafl.remote.remote_struct import RemoteStruct
-from substrafl.simulation import SimulationIntermediateStates
-from substrafl.simulation import SimulationPerformances
-from substrafl.simulation import simulate_aggregate_update_states
-from substrafl.simulation import simulate_test_update_states
-from substrafl.simulation import simulate_train_update_states
+from substrafl.simulation import SimuAggregationNode
+from substrafl.simulation import SimuTestDataNode
+from substrafl.simulation import SimuTrainDataNode
 
 logger = logging.getLogger(__name__)
 
@@ -216,37 +212,6 @@ def _get_packages_versions() -> dict:
     }
 
 
-def _preload_data(
-    client: substra.Client,
-    data_manager_key: str,
-    data_sample_keys: List[str],
-) -> Any:
-    """Get the opener from the client using its key, and apply
-    the method `get_data` to the datasamples in order to retrieve them.
-
-    Args:
-        client(substra.Client): A substra client to interact with the Substra platform, in order to retrieve the
-            registered data.
-        data_manager_key(str): key of the registered opener.
-        data_sample_keys(List[str]): keys of the registered datasamples paths.
-
-    Returns:
-        Any: output of the opener's `get_data` method applied on the corresponding datasamples paths.
-    """
-    dataset_info = client.get_dataset(data_manager_key)
-
-    opener_interface = substratools.utils.load_interface_from_module(
-        "opener",
-        interface_class=substratools.Opener,
-        interface_signature=None,
-        path=dataset_info.opener.storage_address,
-    )
-
-    data_sample_paths = [client.get_data_sample(dsk).path for dsk in data_sample_keys]
-
-    return opener_interface.get_data(data_sample_paths)
-
-
 def simulate_experiment(
     *,
     client: substra.Client,
@@ -295,71 +260,60 @@ def simulate_experiment(
             "`remote` client backend type is not supported by `simulate_experiment`."
         )
 
-    train_data_nodes = copy.deepcopy(train_data_nodes)
-    aggregation_node = copy.deepcopy(aggregation_node)
     strategy = copy.deepcopy(strategy)
-    evaluation_strategy = copy.deepcopy(evaluation_strategy)
+    simu_evaluation_strategy = copy.deepcopy(evaluation_strategy)
 
     train_organization_ids = [train_data_node.organization_id for train_data_node in train_data_nodes]
 
     if len(train_organization_ids) != len(set(train_organization_ids)):
         raise ValueError("Training multiple functions on the same organization is not supported right now.")
 
+    simu_train_data_nodes = []
     for train_data_node in train_data_nodes:
-        train_data_node.update_states = types.MethodType(simulate_train_update_states, train_data_node)
+        simu_train_data_node = SimuTrainDataNode(client=client, node=train_data_node, strategy=copy.deepcopy(strategy))
 
-        train_data_node._datasamples = _preload_data(
-            client=client,
-            data_manager_key=train_data_node.data_manager_key,
-            data_sample_keys=train_data_node.data_sample_keys,
-        )
+        simu_train_data_nodes.append(simu_train_data_node)
 
-        train_data_node._strategy = copy.deepcopy(strategy)
-        train_data_node._intermediate_states = SimulationIntermediateStates()
-
-    if evaluation_strategy is not None:
-        _check_evaluation_strategy(evaluation_strategy, num_rounds)
+    if simu_evaluation_strategy is not None:
+        _check_evaluation_strategy(simu_evaluation_strategy, num_rounds)
         # Reset the evaluation strategy
-        evaluation_strategy.restart_rounds()
+        simu_evaluation_strategy.restart_rounds()
 
-        for test_data_node in evaluation_strategy.test_data_nodes:
-            test_data_node.update_states = types.MethodType(simulate_test_update_states, test_data_node)
-
-            test_data_node._datasamples = _preload_data(
-                client=client,
-                data_manager_key=test_data_node.data_manager_key,
-                data_sample_keys=test_data_node.test_data_sample_keys,
+        simu_test_data_nodes = []
+        for test_data_node in simu_evaluation_strategy.test_data_nodes:
+            # We always take the strategy from the first train node, to obtain a model to test.
+            # TODO: search if a train is on the same node, and take the strategy from it if possible
+            simu_test_data_node = SimuTestDataNode(
+                client=client, node=test_data_node, strategy=simu_train_data_nodes[0]._strategy
             )
 
-            # We always take the strategy from the first train node, to obtain a model to test.
-            test_data_node._strategy = train_data_nodes[0]._strategy
-            test_data_node._score = SimulationPerformances()
+            simu_test_data_nodes.append(simu_test_data_node)
 
-    if aggregation_node is not None:
-        aggregation_node.update_states = types.MethodType(simulate_aggregate_update_states, aggregation_node)
+        simu_evaluation_strategy.test_data_nodes = simu_test_data_nodes
 
-        aggregation_node._strategy = copy.deepcopy(strategy)
-        aggregation_node._intermediate_states = SimulationIntermediateStates()
+    simu_aggregation_node = (
+        SimuAggregationNode(node=aggregation_node, strategy=copy.deepcopy(strategy))
+        if aggregation_node is not None
+        else None
+    )
 
     logger.info("Simulating the execution of the compute plan.")
 
     strategy.build_compute_plan(
-        train_data_nodes=train_data_nodes,
-        aggregation_node=aggregation_node,
-        evaluation_strategy=evaluation_strategy,
+        train_data_nodes=simu_train_data_nodes,
+        aggregation_node=simu_aggregation_node,
+        evaluation_strategy=simu_evaluation_strategy,
         num_rounds=num_rounds,
         clean_models=clean_models,
     )
 
-    train_intermediate_states = reduce(
-        add, (train_data_node._intermediate_states for train_data_node in train_data_nodes)
-    )
+    train_intermediate_states = reduce(add, (node._memory for node in simu_train_data_nodes))
 
-    aggregate_intermediate_states = aggregation_node._intermediate_states if aggregation_node is not None else None
+    aggregate_intermediate_states = simu_aggregation_node._memory if simu_aggregation_node is not None else None
 
     performances = (
-        reduce(add, (test_data_node._score for test_data_node in evaluation_strategy.test_data_nodes))
-        if evaluation_strategy is not None
+        reduce(add, (node._memory for node in simu_evaluation_strategy.test_data_nodes))
+        if simu_evaluation_strategy is not None
         else None
     )
