+++ substrafl/simulation/__init__.py
@@ -1,13 +1,9 @@
-from substrafl.simulation.nodes_utils import simulate_aggregate_update_states
-from substrafl.simulation.nodes_utils import simulate_test_update_states
-from substrafl.simulation.nodes_utils import simulate_train_update_states
-from substrafl.simulation.schemas import SimulationIntermediateStates
-from substrafl.simulation.schemas import SimulationPerformances
+from substrafl.simulation.nodes import SimuAggregationNode
+from substrafl.simulation.nodes import SimuTestDataNode
+from substrafl.simulation.nodes import SimuTrainDataNode
 
 __all__ = [
-    "SimulationIntermediateStates",
-    "SimulationPerformances",
-    "simulate_aggregate_update_states",
-    "simulate_test_update_states",
-    "simulate_train_update_states",
+    "SimuTrainDataNode",
+    "SimuTestDataNode",
+    "SimuAggregationNode",
 ]
